+++ server/app/exportfeatures.js
@@ -146,10 +146,12 @@ this.ExportFeatures = (function() {
       libs: project.libs,
       tabs: project.tabs,
       plugins: project.plugins,
+      libraries: project.libraries,
       date_created: project.date_created,
       last_modified: project.last_modified,
       first_published: project.first_published,
-      files: project.files
+      files: project.files,
+      description: project.description
     };
   };