+++ static/js/project/project.js
@@ -31,6 +31,7 @@ this.Project = (function() {
     this.users = data.users;
     this.tabs = data.tabs;
     this.plugins = data.plugins;
+    this.libraries = data.libraries;
     this.file_types = ["source", "sprite", "map", "asset", "sound", "music"];
     ref = this.file_types;
     for (k = 0, len1 = ref.length; k < len1; k++) {